Changed from roundf to round.
git-svn-id: svn://svn.code.sf.net/p/peano/code/trunk/src@1176 aa977e5b-e928-4c84-aa66-c90e40304b12

+++ tarch/la/ScalarOperations.cpp
@@ -78,10 +78,10 @@ int tarch::la::sign (double number, double tolerance) {
 
 
 int tarch::la::round(double value) {
-  return static_cast<int>(std::roundf(value));
+  return static_cast<int>(std::round(value));
 }
 
 
 int tarch::la::round(float value) {
-  return static_cast<int>(std::roundf(value));
+  return static_cast<int>(std::round(value));
 }
